# api/openapi.yaml
openapi: 3.1.0
info:
  title: LegalMind API
  description: |
    API for LegalMind - Jurisdiction-specific legal template generation with paywall,
    guided questionnaire flow, and attorney referral system.

    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.

    ## Rate Limiting
    - Authenticated: 100 requests/minute
    - Unauthenticated: 10 requests/minute

    ## Errors
    Standard HTTP status codes. Error responses follow format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable message",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: LegalMind Support
    email: support@legalmind.tech
  license:
    name: Proprietary
servers:
  - url: https://api.legalmind.tech/v1
    description: Production
  - url: https://api-staging.legalmind.tech/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
...
paths:
...
  /templates:
    get:
      tags: [Templates]
      summary: List templates
      description: Returns template metadata only. No clause content exposed.
      security:
        - BearerAuth: []
      parameters:
        - name: pack
          in: query
          schema:
            type: string
            enum: [fundraising, saas]
        - name: jurisdiction
          in: query
          schema:
            $ref: '#/components/schemas/JurisdictionCode'
        - name: access_level
          in: query
          schema:
            $ref: '#/components/schemas/AccessLevel'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/TemplateMeta'
  /templates/{template_code}:
    get:
      tags: [Templates]
      summary: Get template metadata
      description: Returns metadata only. Clause content not exposed via API.
      security:
        - BearerAuth: []
      parameters:
        - name: template_code
          in: path
          required: true
          schema:
            type: string
          example: SAFE_PM_V1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateMeta'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===== QUESTIONNAIRES =====
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/me/usage:
    get:
      tags: [Users]
      summary: Get usage statistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Usage stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  generation_count:
                    type: integer
                  last_generation:
                    type: string
                    format: date-time
                  templates_used:
                    type: array
                    items:
                      type: object
                      properties:
                        template_code:
                          type: string
                        count:
                          type: integer

  # ===== ADMIN =====
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from authentication flow

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Partner API key for webhooks
  schemas:
...
    JurisdictionCode:
      type: string
      enum:
        - GLOBAL-EN
        - UK
        - US-DE
        - DE
        - FR
        - CZ
      description: |
        Jurisdiction codes:
        - GLOBAL-EN: Global English baseline
        - UK: United Kingdom
        - US-DE: United States (Delaware)
        - DE: Germany
        - FR: France
        - CZ: Czech Republic
    AccessLevel:
      type: string
      enum: [starter, pro, scale]
      description: Subscription tier
    Plan:
      type: string
      enum: [free, starter, pro, scale]

    # ===== CORE MODELS =====
...
    TemplateMeta:
      type: object
      description: Template metadata only - no clause content
      properties:
        template_code:
          type: string
          example: SAFE_PM_V1
        name:
          type: string
          example: SAFE Agreement (Post-Money)
        pack:
          type: string
        version:
          type: string
        supported_jurisdictions:
          type: array
          items:
            $ref: '#/components/schemas/JurisdictionCode'
        access_level:
          $ref: '#/components/schemas/AccessLevel'
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        usage_count:
          type: integer
        is_featured:
          type: boolean
...
    User:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        plan:
          $ref: '#/components/schemas/Plan'
        jurisdictions_allowed:
          type: array
          items:
            $ref: '#/components/schemas/JurisdictionCode'
        generation_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: FORBIDDEN
            message:
              type: string
              example: Subscription required to access this resource
            details:
              type: object
  responses:
...
    Forbidden:
      description: Forbidden - insufficient permissions or subscription required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            subscription_required:
              value:
                error:
                  code: SUBSCRIPTION_REQUIRED
                  message: This template requires an active Pro subscription
                  details:
                    needed_plan: pro
                    current_plan: free
                    upgrade_url: /purchase
            jurisdiction_not_allowed:
              value:
                error:
                  code: JURISDICTION_NOT_ALLOWED
                  message: Your subscription does not include access to this jurisdiction
                  details:
                    needed_jurisdiction: DE
                    allowed_jurisdictions: ["GLOBAL-EN", "UK"]
                    upgrade_url: /purchase
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'